+++ Backend/findyourngo/type_variables.py
@@ -1,4 +1,4 @@
 # put all your custom types here
 
-TWScore = int
+TWScore = float
